Damn I did a lot of work on this without commiting.
May I forever live in shame. This generates a live-updating site with fake data
that allows for changing of the target temperature. It also generates SVG charts
using vdobler's charts library. The dependencies in the README are a subset of
the actual dependencies, I forgot to keep track until after I had started.